# api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: FluentWork app-server
  version: 0.3.8
  description: Contracts for guest auth, merge, sessions, second-wave review, B17 TTS, B19 B7 hit injection, B22 drill, A4 privacy, B18 review eval, B24 session history, B21 materials, and B23 topic cards.
servers:
  - url: /api/v1
paths:
...
  /account/data:
    delete:
      summary: Soft-delete the authenticated user's data (A4)
      description: |
        Requires confirmation_code `DELETE-MY-DATA`. Soft-deletes users,
        phrase_blocks, and practice_sessions; hard-deletes drill_records;
        anonymizes ai_cost_logs. Idempotent on a second call. Wrong confirmation
        returns 422 FAILED_PRECONDITION. Per 48 §1.1.6.
      operationId: deleteAccountData
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountDataRequest"
      responses:
        "200":
          description: Cascade counts and backup purge deadline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountDataResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /sessions:
    get:
      summary: List the caller's practice sessions
      operationId: listSessions
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Cursor page of sessions (newest first)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListPage"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a practice session and issue a one-time WSS ticket
      operationId: createSession
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                material_id:
                  type: string
                  nullable: true
                  maxLength: 36
                scene_type:
                  type: string
                  maxLength: 64
                  default: demo
      responses:
        "200":
          description: Session id, WSS URL, and ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /materials:
    post:
      summary: Create a material and enqueue phrase-block refine
      operationId: createMaterial
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                kind:
                  type: string
                  enum: [paste, sentence, url]
                  default: paste
                content:
                  type: string
                  maxLength: 5000
      responses:
        "202":
          description: Material queued for refine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMaterialResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /topic-cards:
    get:
      summary: List today's topic cards
      operationId: listTopicCards
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Today's cards (possibly empty before generation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicCardList"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /internal/v1/tts/synthesize:
    post:
      servers:
        - url: /
      summary: Synthesize speech with Volc streaming TTS and duplex fallback
      description: |
        Internal endpoint (X-Internal-Token). Primary provider is Volc HTTP/2
        unidirectional TTS. After three consecutive HTTP 5xx responses, traffic
        switches to the duplex TTS fallback and increments
        tts_fallback_triggered_total{from="volc_streaming",to="volc_duplex"}
        on GET /metrics.
      operationId: synthesizeTTS
      security:
        - internalToken: []
      parameters:
        - name: voice_id
          in: query
          required: false
          schema:
            type: string
          description: Overrides JSON voice_id. D-2 catalog or a raw Volc speaker id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TTSSynthesizeRequest"
      responses:
        "200":
          description: Concatenated audio chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TTSSynthesizeResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /internal/v1/voicegateway/hits:
    post:
      servers:
        - url: /
      summary: Record B7 phrase-block hits from the voice gateway
      description: |
        Internal endpoint (X-Internal-Token). UPSERTs phrase_block_uses for
        (session_id, turn_id, block_id) and increments phrase_blocks.total_uses
        only on first insert. Duplicate reports of the same hit do not double-count.
      operationId: recordVoiceGatewayHits
      security:
        - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceGatewayHitsRequest"
      responses:
        "200":
          description: Number of hits written to the ledger (including UPSERT updates)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceGatewayHitsResponse"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  schemas:
...
    DeleteAccountDataRequest:
      type: object
      required: [confirmation_code]
      properties:
        confirmation_code:
          type: string
          description: Must be DELETE-MY-DATA
    DeleteAccountDataResponse:
      type: object
      required: [cascaded, backup_purge_at, already_deleted]
      properties:
        cascaded:
          type: object
          additionalProperties: { type: integer }
        backup_purge_at: { type: string, format: date-time }
        already_deleted: { type: boolean }
...
    CreateSessionResponse:
      type: object
      required: [session_id, wss_url, ticket, ticket_expires_in, ticket_expires_at, scene_type, status]
      properties:
        session_id: { type: string }
        wss_url: { type: string }
        ticket: { type: string }
        ticket_expires_in: { type: integer }
        ticket_expires_at: { type: string, format: date-time }
        scene_type: { type: string }
        status: { type: string }
    SessionListPage:
      type: object
      required: [items, size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SessionListItem"
        next_cursor:
          type: string
        size: { type: integer }
    SessionListItem:
      type: object
      required: [session_id, scene_type, status, started_at, duration_sec]
      properties:
        session_id: { type: string }
        scene_type: { type: string }
        status: { type: string }
        started_at: { type: string, format: date-time }
        duration_sec: { type: integer }
        material_id: { type: string, nullable: true }
...
    TTSSynthesizeRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
        voice_id:
          type: string
          description: D-2 id (zh_male_tech_01, en_female_professional, en_male_narrator, en_female_clear) or a raw Volc speaker
    TTSSynthesizeResponse:
      type: object
      required: [voice_id, chunks, audio_base64, used_fallback]
      properties:
        voice_id: { type: string }
        chunks: { type: integer }
        audio_base64: { type: string }
        used_fallback:
          type: boolean
          description: true after Manager switched from volc_streaming to volc_duplex
    VoiceGatewayHitsRequest:
      type: object
      required: [user_id, session_id, turn_id, hits]
      properties:
        user_id: { type: string }
        session_id: { type: string }
        turn_id: { type: string }
        hits:
          type: array
          items:
            $ref: "#/components/schemas/VoiceGatewayHit"
    VoiceGatewayHit:
      type: object
      required: [block_id, detected_at_ms]
      properties:
        block_id: { type: string }
        detected_at_ms: { type: integer, format: int64 }
        turn_id:
          type: string
          description: Optional per-hit turn; defaults to the request turn_id
    VoiceGatewayHitsResponse:
      type: object
      required: [recorded_count]
      properties:
        recorded_count: { type: integer }
...
    CreateMaterialResponse:
      type: object
      required: [material_id, refine_status]
      properties:
        material_id: { type: string }
        refine_status:
          type: string
          enum: [queued, processing, ready, failed]
...
    TopicCardList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TopicCard"
    TopicCard:
      type: object
      required: [id, for_date, title, prompt_en, prompt_zh, card_type, seed_tags, valid_until, created_at, updated_at]
      properties:
        id: { type: string }
        for_date: { type: string, format: date-time }
        title: { type: string }
        prompt_en: { type: string }
        prompt_zh: { type: string }
        card_type:
          type: string
          enum: [warmup, practice, stretch]
        seed_tags:
          type: array
          items: { type: string }
        valid_until: { type: string, format: date-time }
        checked_in_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }